RTCConfiguration's default value for sdp_semantics is about to change from kPlanB to kUnifiedPlan. Because of this, all uses of RTCConfiguration has to be updated to explicitly set the sdp_semantics field.
The NO-OP version of this CL would be to set it to kPlanB and
file a buganizer issue to track migration from kPlanB to
kUnifiedPlan, but because nearby's use cases only uses data
channels (i.e. it does not use WebRTC for audio and video)
nearby should not care about which SDP semantics is used, and as
such this CL goes ahead and updates it to kUnifiedPlan.

For more info, see go/plan-b-removal-and-intro-to-unified-plan
and b/208964426.

PiperOrigin-RevId: 417496429