Implemented Windows implementation compliant Linux `http_loader`
The `http_loader` class takes a `WebRequest` with information to send a
web request, and sends it. This uses `libcurl`, which is present from
installation on most major Linux distrobutions (e.g. Fedora, Ubuntu,
PopOS...).

If libcurl is not present on the target system, it will need to be
installed.