[Nearby Connections] Use borrowable model for ClientProxy
Instead of passing around a raw pointer to the ClientProxy, pass around a borrowable reference to the ClientProxy, which will fail safely if the ClientProxy has been destroyed. To use ClientProxy, callers ask to Borrow() an instance of the ClientProxy and check if the borrowed instance is valid before continuing.

Borrow() will block if another thread is already borrowing the object, so this CL introduces FinishingBorrowing() to allow callers to release their borrowed instance and not block.

PiperOrigin-RevId: 560233682