Remove pointer-based overload of RandBytes()
The pointer-based overload is going away in Chromium base: https://chromium-review.googlesource.com/c/chromium/src/+/5514816

Add `base/containers/span.h` to the Chromium build shims, and use it from
`internal/crypto_cros/random.h`. This makes the signature match and also provides
all the base span helpers that are needed to work with spans nicely.

Then we use those helpers to convert calls to the span-based `RandBytes()`.

There is previously some confusion in the Chromium shims as they have the same names
and namespaces as Chromium but not always the same signatures. And the shims are
sometimes included even in a Chromium build, which produces ODR violations. To resolve
this we:
- Split random.h and random.cc into a separate `random_chromuim` build target.
- Introduce new base headers from chromium into a `base_chromium` build target.
- Introduce new build headers from chromium into a `build_chromium` build target.
- In every chromium shim header add an `#error` if included inside the chromium build,
  which is when `NEARBY_CHROMIUM` is defined.

Then, we fix a few include sites of random.h to only include the shim when not in the
Chromium build. As we find the `#error` hit in the Chromium build now.